Add trash functionality to product and category manipulation in admin dashboard.

+++ resources/views/admin/categories/index.blade.php
@@ -4,7 +4,9 @@
     <div class="container mt-5">
         <h1 class="mt-5 rounded-pill bg-dark text-center p-3 text-light"> Categories</h1>
         <div class="mt-5">
-            <a href="{{ route('admin.categories.create') }}" class="btn btn-primary btn-lg" role="button">Add Category</a>
+            <a href="{{ route('admin.categories.create') }}" class="btn btn-primary btn-lg" role="button">Add
+                Category</a>
+            <a href="{{ route('admin.categories.index',['only_trash']) }}" class="btn btn-warning btn-lg" role="button">Trash</a>
         </div>
         <div class="mt-3">
             <div>
@@ -12,9 +14,9 @@
                     <thead class="thead-dark">
                     <tr>
                         <th scope="col"><h4> #</h4></th>
-                        <th scope="col"> <h4>Name</h4></th>
-                        <th scope="col"> <h4>Slug</h4></th>
-                        <th scope="col"> <h4>Description</h4></th>
+                        <th scope="col"><h4>Name</h4></th>
+                        <th scope="col"><h4>Slug</h4></th>
+                        <th scope="col"><h4>Description</h4></th>
                         <th scope="col"></th>
                     </tr>
                     </thead>
@@ -32,18 +34,30 @@
                             <td>
                                 <div class="row">
                                     <div class="form-inline">
-                                        <form action="{{route('admin.categories.edit', ['category'=>$category->slug])}}"
-                                              method="get">@csrf
-                                            <button class="btn btn-success btn-sm" type="submit"><i
-                                                    class="material-icons">edit</i></button>
-                                        </form>
-                                        <form action="{{ route('admin.categories.delete', ['category'=>$category->slug]) }}"
-                                              method="post">@method('DELETE')@csrf
-                                            <button class="btn btn-warning btn-sm m-2" type="submit"><i
-                                                    class="material-icons">delete</i></button>
-                                        </form>
-                                        <form action="{{ route('admin.categories.forceDelete', ['category'=>$category->slug]) }}"
-                                              method="post">@method('DELETE')@csrf
+                                        @if(!$category->deleted_at)
+                                            <form
+                                                action="{{route('admin.categories.edit', ['category'=>$category->slug])}}"
+                                                method="get">@csrf
+                                                <button class="btn btn-success btn-sm" type="submit"><i
+                                                        class="material-icons">edit</i></button>
+                                            </form>
+                                            <form
+                                                action="{{ route('admin.categories.delete', ['category'=>$category->slug]) }}"
+                                                method="post">@method('DELETE')@csrf
+                                                <button class="btn btn-warning btn-sm m-2" type="submit"><i
+                                                        class="material-icons">delete</i></button>
+                                            </form>
+                                        @else
+                                            <form
+                                                action="{{ route('admin.categories.restore', ['category_slug'=>$category->slug]) }}"
+                                            method="post">@csrf
+                                                <button class="btn btn-primary btn-sm m-2" type="submit"><i
+                                                        class="material-icons">restore</i></button>
+                                            </form>
+                                        @endif
+                                        <form
+                                            action="{{ route('admin.categories.forceDelete', ['category_slug'=>$category->slug]) }}"
+                                            method="post">@method('DELETE')@csrf
                                             <button class="btn btn-danger btn-sm" type="submit"><i
                                                     class="material-icons">delete_forever</i></button>
                                         </form>
